Split the OVS plugin and agent properly.
The OVS plugin operates from within the quantum server and should
a. only affect configuration and b. only that configuration necessary
for getting the server to run.  Since the OVS plugin talks to agents
about networks (as named in network_vlan_ranges and bridge_mappings
on either side) this means that network_vlan_ranges belongs in the
plugin and bridge_mappings in the agent block.

All naming and manipulation of interfaces and bridges moves to the
agent - the plugin code is abstracted away from this level.

All discussion about how to allocate tags (vlan or tunnel) moves to
the plugin - the agent is told what to do with them and never
chooses or checks them itself.

This code is untested, I've merely attempted to make the split clear.
It has also drastically changed the calling conventions for the two
blocks.