Minor changes to quantum::agents::ovs
This patch contains two minor edits to the
quantum::agents::ovs code.

1. allow setting of the firewall_driver.

This code ensures that the firewall driver can be
either set to its default of IptablesFirewallDriver
or unset (which we believe is required in the case
of compute nodes with gre enabled)

2. explicitly remove tunneling configuration when
it is not being used. This is part of an ongoing
attempt to clean up configuration that are not
currently in use.

This patch also adds tests to verify the proposed
code.

Change-Id: I372889d1c8821605d60c91c1e244d1a4c83fcc5e