Drop libvirt qemu configuration
This configuration fix with regards to cgroup ACL should go to the
puppet-nova module instead. With latest OpenStack Network architecture
there are valid setups to use this module dedicated for the network
part without any compute service, like libvirt, runnin on that node.

This got introduced due to this bug:
https://bugs.launchpad.net/openstack-cisco/+bug/1086255

I guess the issue is/was that /dev/net/tun access is missing by default
for some distribution. Which is required if OVS is used and needs to
be configured by libvirt.

But such fix should go to puppet-nova or a libvirt puppet module instead.

Change-Id: I3b4f0ec03e3f2e2041598ee2c6086ef8157518da