Don't rely on Array.to_s which differs between Ruby 1.8 and 1.9
The should contain_quantum... call is relying on the stringification
of params bridge_mappings/bridge_uplinks arrays for a successful match.
With ruby 1.8, this currently works by accident since the array only
has 1 value and is stringified to something like 'default:br-ex'.
With ruby 1.9, it is stringified to '[default:br-ex]' causing the
match (and therefor the tests) to fail.

This patch bypasses the differing to_s implementations by just
calling join(',')

Change-Id: I6ab9fd6a04cdee652c86512d871560effd7b08a5