Remove native type prefetches
The network, router, subnet and router interface
providers attempt to prefetch the configuration,
which can result in empty or stale values being
loaded as quantum configuration. The former will
result in errors stating the quantum.conf file
is incomplete, while the latter will often
result in authentication errors.

This patch removes the prefetching in the child
classes and explicitly sets the conf to nil in
the parent quantum provider so that it will be
loaded at the correct time.

Change-Id: Ia2dbc8b687498be96d09c9286fb3f07e0c5b8f96
bug 1197588