Support Cisco quantum plugin
This patch adds native types for the config files
needed to enable the cisco quantum plugin as well
as a class that will create these config files.

Users will be able to select their plugin
of choice by setting the core_plugin on the
openstack::controller, and then creating
the quantum::plugins::cisco class separately.

Change-Id: I8c131ebdbab149f589c84e52be8f810e39624941