Puppet Forge 2.1.1 Release
Prior to this commit, the Modulefile incorrectly listed the
puppet-quantum module version as 0.2.2. This patch updates
the version to 2.1.1. As the repository has already been
given a release tag of 2.1.0, this is a candidate for a new
release with a minor bug fix version bump. No 2.1.0 module
will be available.

Change-Id: Id94800e318e4560b6ba5555f43f849d8df028f5a
(cherry picked from commit 2d55ecc9218b3796a5f012f752f0004016c10edf)