Scrap nonsense provider setting requirements
Currently, if any provider setting is defined, it's enforced that
the other settings are defined as well. This is wrong and in some
cases even contradicts with Neutron.

Different settings need to be either present or absent depending on
the various network types. It would be a huge overkill to validate
all the possible combinations, particularly since relying on
Neutron to do the proper checks is good enough.

Change-Id: I36ad13da6d796d79e5d0050a13f1766455b478be