Fix improper test for tunnel_types param
When VXLAN support was added to the ovs agent manifest, a new
parameter called tunnel_types was also added with the default
value being an empty array.  However, later in the code there's
an if loop that tests truth--and an empty array returns true.
This causes a line of the form:

tunnel_types =

to show up in /etc/neutron/plugins/openvswitch/ovs_neutron_plugin.ini
if the default value is used, which causes OVS to refuse to start.
This patch corrects the problem by properly testing the array.

Change-Id: I06ad6249ab7803676e49c026b4692f06ade269ed
Closes-Bug: #1281684