Fix FWaaS race condition
The current FWaaS manifest suffers from a race condition: it may
attempt to configure the FWaaS service before the Neutron packages
on which it depends has been installed, and before the
/etc/neutron directory exists.  This leads to errors on the first
catalog run, and the FWaaS service doesn't start.  The errors
go away on the second catalog run since by then the requisite
directory exists.  This patch adds ordering such that the packages
on which the FWaaS service relies are installed before the FWaaS
service is configured.

Change-Id: I174cf06b9e8355bfe499b23f390d4078dc670e72
Closes-Bug: #1287554