Don't sync database by default
https://review.openstack.org/#/c/76358 introduced a new sync_db
parameter and set it to true by default.  Unfortunately there
seem to be many cases where this causes problems.  One is when
using the OVS plugin with the [database] section stored in
neutron.conf rather than in the ovs_neutron_plugin.ini file.
In such cases the migration will error out with due to being
unable to connect.  This is particularly problematic for users
in that no changes to the composition layer are necessary to
hit the error since sync_db is set to true by default: to avoid
the problem they need to make composition layer change to explicitly
set it to false.  This patch resolves the problem by setting the
parameter to false by default rather than true.  This enables
users who want the migration to run to do so without impacting
users who haven't been expecting the new behavior.

Change-Id: Id8c116212132b6d75ae14019fc6a65447f226c6d
Closes-Bug: #1288975