Don't sync database by default
https://review.openstack.org/#/c/76358 introduced a new sync_db
parameter and set it to true by default.  Unfortunately there
seem to be many cases where this causes problems.  One is when
using the OVS plugin with the [database] section stored in
neutron.conf rather than in the ovs_neutron_plugin.ini file.
In such cases the migration will error out with due to being
unable to connect.  This is particularly problematic for users
in that no changes to the composition layer are necessary to
hit the error since sync_db is set to true by default: to avoid
the problem they need to make composition layer change to explicitly
set it to false.  This patch resolves the problem by setting the
parameter to false by default rather than true.  This enables
users who want the migration to run to do so without impacting
users who haven't been expecting the new behavior.

This patch is based on 0229d358ae45add73a0a8f4658632b54fd56be12
from master but is modified slightly to account for a change
in the rspec tests to use the havana instead of head during
the dbsync.

Conflicts:

	spec/classes/neutron_server_spec.rb

Change-Id: Id8c116212132b6d75ae14019fc6a65447f226c6d
Closes-Bug: #1288975
(cherry picked from commit 0229d358ae45add73a0a8f4658632b54fd56be12)