Make haproxy package management optional
Puppetlabs-haproxy does not use ensure_resources
and can cause failures when combined with lbaas
agent class.

This patch makes package management of haproxy
optional so that agents::lbaas can play nicely
with that module without requiring a patch there
changing to ensure_resource.

The dependency is fairly safe since:
 - redhat package name is haproxy
 - debian package name is haproxy
 - puppetlabs-haproxy aliases the package name to haproxy

If the names of any of these things get out of sync, this
may lead to errors.

Change-Id: I6e603cf328ade9460e98e391ef01337fcb6ea72b