if log_file enabled, ensure log_dir is not absent
If log_file is specified, that means we should configure log_dir.
Otherwise, when running a command which use debug (like db-sync commands), it will try to write logs in the user directory which is not the best way.

This patch aims to ensure that we configure log_dir instead of remove
the flag like it was before.

Partial-bug #1292619
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>

Change-Id: I94528b42a4865976f74d264facd91cb521c177bd