ml2-plugin: update security group option
In Icehouse, security group configuration is enabled from the plugin and
should be enabled by a boolean value instead of a firewall_driver like
before.
If security group is enable, we should ensure firewall_driver is true or
a specific driver.
If security group is disabled, we should use Noop firewall driver.
Patch in Icehouse: http://git.io/Bb-PHQ

Change-Id: I0d0ca2fb80d21a439c5e096dcd978e956e6662fd
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>