allow neutron::agents::ovs without package install
This alters the service notification and dependency behaviour for
neutron::agents::ovs such that if the package is not installed (i.e.
package_ensure => absent, enabled => false) the class still runs without
producing errors.  This is to allow the generation of config files
without needing to install the package in the first place.

Change-Id: I027313518f778d58e6e19b9b933ff17c0094a8aa
Closes-Bug: #1298146