fix wrong location of report_interval
report_interval is a config in section agent, not section DEFAULT.
It is used by the agents to scheduly report state, and not used by
neutron-server. Missing this config on agent side, leads to flapping
alive status of neutron agents. Therefore report_interval is moved
from class neutron::server to class neutron

Change-Id: I4fa84dabcba8465b297f6dc0eefa1d02a98eafca