Neutron Nova Interactions Support
Recent versions of Neutron are able to provide active notifications to
Nova about changes to port status and data.  This change adds the
neutron::server::notifications class to allow Puppet to configure this
feature when setting up a Neutron server.

neutron.conf requires a tenant UUID for communicating with Nova.  This
changes adds adds a custom type and provider to translate an admin
tenant name to the required UUID.

Co-Authored-By: emilienm <emilien.macchi@enovance.com>
Change-Id: I57af7cb47a12adabed9b1ee65e1c352b3e5fc2de
Closes-bug: 1306694