Add parameter to set veth MTU
In many applications (including in particular NFV), it is useful
to use jumbo frames on veth pairs.  The OVS agent has a setting
enabling the user to set the MTU on veth interfaces, but it is
currently not exposed as a parameter.  This patch exposes the
veth_mtu setting and adds rspec tests accordingly.

Change-Id: I7b211be29a5630874b1c56e79ca80667f96ce425
(cherry picked from commit 3b6abf010a2c14f296044291cd381a6a7a4f92d9)