Add network_device_mtu to neutron.conf
Already a parameter for L3_agent, but that isn't
exactly the right configuration, due to the fact
that this parameter is specifically a L2 feature.

Added in neutron.conf to set the default MTU for all
created networks. Discussion available here:

http://lists.openstack.org/pipermail/openstack-dev/2013-October/017033.html
https://bugs.launchpad.net/openstack-manuals/+bug/1242534

Added test to validate parameter
Fixed mis-spelled parameter in test
Added deprecation warninigs to agents/l3.pp

Closes-Bug: #1329643
Change-Id: Id8544799855c500f0c871681695db502ab4dc87c