Link from ovs config file to plugin.ini on RedHat
All init scripts for neutron-openvswitch-agent from RedHat
packages use the non-ml2 config file path. When using ML2
this causes problems since all config is set under the ML2
plugin directory.

Since this is only a bug when using ML2, but not when using
the monolithic plugin, it may not be fixed in packages.

https://bugzilla.redhat.com/show_bug.cgi?id=1087647

This patch creates a symlink from the old plugin config
path to the new one to mitigate the issue.

Change-Id: I5feb8fc39fa4d67aebb4209d7b44bea38d1a9d5f