Make api_extensions_path a parameter for "neutron"
Currently, the "neutron" class does not support the direct configuration
of the "api_extensions_path" by means of a parameter. However, there is
at least one plugin -- OpenContrail -- which currently requires exactly
this parameter to be set to work. It is hence desireable to be able to
set the parameter directly from within the "neutron" class instead of
having to fiddle with "neutron_config" in local site manifests.

This commit adds the desired functionality and allows the parameter via
"api_extensions_path" to be set directly for the "neutron" class.

Change-Id: I660848100073a3a0067ffb42791bfd80b66c808e