Subscribe neutron db sync to db connection setting
Currently, neutron db sync is requires the db connection to
be set, but is only triggered by changes in the neutron package.
Since the neutron package, containing the neutron-server binary
is a dependency of other neutron services, this means that if
any other neutron service is installed before neutron-server, the
db will not be synced.

This patch modifies the exec resource to subscribe to the db connection
string, handling both the above case and also the case where the
db connection string is changed to point to a new database for
whatever reason and must be re-synced.

Change-Id: I6ce3e31f048dfc47cd39bbbbea9de0934f015ba4