Correct docs on format for neutron::policy data
Correct the format of the $policies parameter to neutron::policy.
That parameter is passed directly to create_resources, so it needs
to be in a format acceptable to that.

Just updating docs here as anyone already using the policy classes
are likely not using the (incorrectly) documented format, anyway.

Change-Id: Ib69e2e421b216d6087ad47c198067667db89cc33
Closes-bug: 1409897
(cherry picked from commit 6d2aa10d8f8ed2dc05ee300e067e431d02d2209e)