Fix catalog compilation when not configuring endpoint
Previously when you set $configure_endpoint to false and did not have a
keystone_endpoint resource with the same name as is defined by
neutron::keystone::auth, the catalog would not be compiled. This was
because a relationship was being established where one resource in the
relationship didn't exist.

This changes this so that the relationship involving the Keystone
endpoint is only defined if $configure_endpoint is set to true.

Change-Id: I710da4811f456410239efb1487e3407a42686a2b
Closes-Bug: 1368686