Add MidoNet plugin manifest
Define the resources to configure MidoNet plugin with Neutron. Is worth
say that, since MidoNet plugin is not in official repos, this manifest
does not install the packages and delegates to an orchestration tool or
a wrapper manifest to do so.

Is is planned to add package declaration after Kilo release.

Change-Id: Ib5f73729f314a40197b4d1c75ec33fd246df83d9
(cherry picked from commit 5e034e2af7ecb31cfcb758c7f43f47e46ce5677a)