provide explicit defaults for boolean values
When creating a neutron_network resource without explicitly setting the
router_external parameter, the network would be created with
router:external=True, which would result in a network to which it was
not possible to attach nova instances.

This change sets explicit defaults for all boolean options in the
neutron_network type.

Closes-bug: 1436133
Change-Id: Id0f950da791c65ba9034046177870f31f6f57c46