Reduce neutron API timeout to 10 seconds
If the neutron server is down, 120 seconds is far too long to wait for
puppet to fail. This patch reduces the timeout to 10 seconds because if
the neutron server is flaky, but not down, 10 seconds is a reasonable
amount of time to retry connecting.

Change-Id: I7bd694d1e5ee3477f7ea862965f7f3368903abde
(cherry picked from commit 26f1ae699667864739ce7fbe74c8257ff3fba720)