Fix neutron file_line dependency
Without this patch, on Ubuntu the neutron module will sometimes try to
apply the NEUTRON_PLUGIN_CONFIG file line change to the file
/etc/default/neutron-server before it exists. On Ubuntu, this file is
managed by the neutron-server package. Since we otherwise do not need
to manage that file on its own, we add a dependency on the package
resource to make sure the file will be in place before trying to modify
it with file_line.

Change-Id: Ic6905fe4240934f526e4b0d21bee695c8d93f28c
(cherry picked from commit a731408167847854ed39decf1b00ae5f698bcda6)