Make retry message more descriptive
When neutron fails the neutron client is often very helpful in
describing the issue. This message should not be hidden in a debug
statement, we should be able to see it easily since it is often easy to
fix. This patch removes the debug output and puts the error message in
the notice.

Change-Id: Ic7769ba463bca22af8c5b1ebc9c43e2b25a0ca99