Fix parsing of network gateway id for router
The neutron router-show command with shell formating outputs the
external_gateway_info with literal backslashes to escape the double
quotes. The provider must then match against the literal slashes when
extracting the gateway id. If this is not done, the provider cannot
detect the gateway id from the external_gateway_info property and will
always attempt to change the gateway name from '' to the desired name.

Change-Id: I85a18ed08886cebd14f61b65b3681a75a8d41741