Fixed issue with rabbit_ha_queues parameter
We need to configure rabbit_ha_queues = True only if we have
more than 1 RabbitMQ host. In we have one RabbitMQ host we need to
set rabbit_ha_queues = False.
We also need to use one pattern for configuration of this parameter
for different OpenStack components.

Also fixed issue with rabbit_hosts parameter, because we should initialize
it with array, not with string or boolean value.

Change-Id: Ia610241351074feccfbd54fab270ab27caa8a34b
Closes-Bug: #1470054