Make sure Neutron_config is applied before Neutron_network
When creating a Neutron_network, the provider will read the actual
configuration file. If Neutron_config has not be applied before, the
default configuration will be used and probably Keystone
authentification will fail.

So this patch aims to make sure neutron.conf is filled before trying to
create a network.

Depends-on: Ibd90e5e159f9a1fbff3990241795894e0bdca3cc
Change-Id: Ia9141cb387a98c70cc639db724da4af4b3289d8a