acceptance: disable lbaas service plugin
Neutron packaging seems having issues when LBaaS is enabled in
service_plugins Neutron parameter.

As a temporary workaround, we disable it so we can unblock our CI, but
keep LBaaS service installed & working, so we continue the class
testing.

Once packaging is fixed, we will enable the service plugin again.

It also drop the device_driver parameter which is already the default
one in puppet-neutron.

Partial-bug: #1490990
Change-Id: Ibd90e5e159f9a1fbff3990241795894e0bdca3cc