Allow greater flexibility in the kind of parameter passed to ml2.pp
Currently the user is forced to pass an array for some values such as
$network_vlan_ranges when an actual string would be definitly a correct
value.

This was done because later on the processing, the code relies on
function that expects an array, yes join() we are talking about you.

The use of any2array() can give us the flexibility to receive either a
string or an array.

The affected variables are the following :

 * type_drivers
 * tenant_network_types
 * mechanism_drivers
 * flat_networks
 * network_vlan_ranges
 * tunnel_id_ranges
 * vxlan_group
 * vni_ranges
 * supported_pci_vendor_devs

Change-Id: I0496e2f855f251a77329163845f87efe1b93cbb1