Make vswitch optional for ovs agent configuration
For some complex configurations it may not be ideal to try and leverage
the vswitch::ovs class to manage the software and ovs bridges and ports.
This change adds a flag to the ml2 ovs agent configuration to allow the
vswitch configuration to be skipped as part of the agent installation
and configuration. If new parameter manage_ovs is set to false, it will
skip the inclusion of vswitch::ovs and attempts to configure the ovs
bridges/ports based on the bridge_mappings paramters. This allows a user
to manage the openvswitch installation and configuration elsewhere.

Change-Id: Iee0edf08d9a96580bec2b80274102c1da1a70ed0