Simplify rpc_backend parameter
Currently in init.pp, the value of rpc_backend must be
'neutron.openstack.common.rpc.impl_kombu' for the rabbitmq
variables to be set. Since the official guide suggests
using the alias 'rabbit' for the rpc_backend, it should
also be an accepted value.

This patch maintains backwards compatibility with the previous
default value in case it is explicitly used.

Change-Id: I05aa69bac55e06e90ab9e1d7c4f5e940853ae709
Closes-bug: #1431192