Add custom "neutron" command override
This patch adds a custom "neuton" command
implementation. It outputs only stdout of
the python cli tool and drops the stderr.
If error have happened both stdout and
stderr go to the exception message.

This can fix the situation when neutron
tool outputs several non-critical warnings
and they break the output parser in the
provider.