Add workaround for v2 tenant names
The change (https://review.openstack.org/#/c/226919/) to the naming
convention for keystone tenants has introduced an issue with some
neutron providers as they are no longer able to find the v2 named
tenants since the name returned from the keystone provider now
also includes the domain. This change adds a check to see if the name
matches the Default domain applied by default until support for a
resource without the domain set is removed.

Change-Id: I4c5b6b2b663a1d08005369fa755ac03b0603102b
Closes-Bug: #1515787