Neutron parses the wrong json revert to shell/csv.
The neutron network provider is parsing the wrong json format leading
to an error when you're unlucky. Affected providers:
 - neutron_port
 - neutron_subnet
 - neutron_router
 - neutron_network

The neutron provider parses the json of the cliff-tablib's output while
the official one is the cliff' output[1].

As discussed in the ML[2] this patch implements a simple revert of the
json change.

Revert "Use json output instead plain-text"

This reverts commit 79f0e636c07b80b2c6699d934b42641bfef4352b[3]
 and part of commit 791a0f146b28544ec3d70c0ab2a950c4d5ca9f98[4]

[1]: https://bugs.launchpad.net/python-neutronclient/+bug/1529914
[2]: http://lists.openstack.org/pipermail/openstack-dev/2015-December/083172.html
[3]: https://review.openstack.org/#/c/238156/
[4]: https://review.openstack.org/#/c/261418/

Change-Id: I93263256eb9d83544910ca4055bb3e9a463645e3
Closes-bug: 1530163