Make neutron::server inherit from neutron::params
This change updates the neutron::server class to inherit from
neutron::params because we are using ::neutron::params::server_service
as the default for the service name. Without this inherit, a warning may
be raised because the params class may not be loaded prior to trying to
set the default value for the service name.

Change-Id: If8f06b8ab5bf50f56090646f1bc829b231232525
Closes-Bug: #1531249