Do not manage network, subnet, router resources if prefetch fails
If prefetch of neutron_network, neutron_subnet, or neutron_router
resources fails (because of Neutron API timeouts or failures, for
example) then the provider will believe the resources don't exist
and will attempt to create them.  This can cause duplicate
resources in Neutron.

This patch records when the prefetch fails, and if so, shunts
any of the resource management methods to fail().

Change-Id: I612847bfd23ac735583ab228d8a9396f52f0ef14
Closes-bug: 1532284