Support the LibreSwan driver in Red Hat platforms
Currently, the libreswan package is supported by puppet-neutron as
part of the openswan driver. However, there is a specific libreswan
driver, which is required when using it (see
https://bugs.launchpad.net/neutron/+bug/1444017 for details).

Adding support for the LibreSwan driver in puppet-neutron, only for
Red Hat platforms as it does not seem to be supported in Debian/Ubuntu.

Change-Id: I9628e485d301734f9d739c1d03d6da9f7887c61d
(cherry picked from commit f6c48bb76c95fe09e05f24fc4437d20fdb5230d2)