Link linuxbridge agent to rabbitmq changes
The linuxbridge agent on compute nodes needs to be able to talk to
rabbitmq. If the rabbitmq settings change, i.e. moves ports or hosts,
all the services that talk on rabbitmq need to know about it and
refresh their configs. This patch makes sure the linuxbridge agent
refreshes when the rabbitmq_hosts config changes in neutron.conf.

Change-Id: Ic8e0e48d303cc3ed7b51560352a6585ccd977b7f