change default value for lock_path
oslo.concurrency does not provide default value for lock_path.
So if you don't provide it (default), neutron-server will send lot of
ERRORS that cause issues when dealing with networks management.

Even though the default is None in OpenStack, we want to change this
poor user experience to a successful deployment by default.

Closes-Bug: #1549972
Change-Id: I8a5ba016fc788bda1cd3f3d7666a78ba219d3ed3