Added deprecation warning for lbaas agent.
Using the parameter to install the lbaas agent on the neutron
server should be avoided. If the user wants to install the lbaas
agent on the neutron server they should be using the lbaas agent
class. This change adds a deprecation warning with information
to the user on the proper way of installing the lbaas agent.

I have updated the comment for the parameter as the lbaas agent
is not required by neutron server to enable the lbaas service plugin.

Change-Id: I23bdb1385e9ff3e7218bf4c05f229cce4caf2504