Fix unit tests
The unit tests were broken by a change to openstacklib
Id09c3358c5843510e6a2a8c0e2d4aeb3607e098b which renamed some of the
resources in the openstacklib classes. The problem is that our tests in
this module should not have been testing what occurs within openstacklib
so this fix updates the tests to only check what we are doing within
this module.

Change-Id: Idb09a2ecc467891942e253de79980b142235a838
Depends-On: I9d535ab38afea852559df2b3073bd4b74a2a3947