Added service authentication options to LBaaS
By default, LBaaS is configured to authenticate to the local instance
of Keystone running at 127.0.0.1:5000.  In large deployments, it will
likely not be there and it will fail to create anything.

This patch adds the options to allow controlling these options in order
to be able to successfully create load balancers.

Change-Id: I76c941235b0177ec5dffe100b0866fab9045f209