Adds the ability to specify ML2 ODL Features
This is required in order to override the negotiated default features
with OpenDaylight.  More specifically this is required so that we can
override the features list not to include operational-port-status to
fix the following bug.

Partial-Bug: 1718508

Change-Id: I7eb752ad692e5522051f8393376890fcac9a09fe
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 2417be6fae4df4003be0471a405d09a653d29f5d)