Set agent debug config defaults to service default
A few of the neutron agents had a default of 'false' for the debug
flag. This interferes with the common practice of enabling debug through
a common configuration file (e.g. /etc/neutron.conf). This patch changes
the default to ::os_service_default allowing it to either use the
service default, values set in some common configuration or explicitly.

Change-Id: I7d7a1e4776861dc8d94a143edc0df76c3f963ef8
Closes-Bug: #1733071