Corrects ODL-OVS vhostusermode default and docs
The vhostuser_mode parameter was incorrectly defined as the
client/server mode of OVS, when actually it is the configuration of the
VIF port/QEMU mode.  Therefore a value of 'client' will actually
configure OVS to be in 'server' mode.  This type of port is deprecated
and the correct value should be 'server' (where QEMU creates the port).

Partial-Bug: 1762473

Change-Id: Idda82cae48b11eae1c79c6bbdd85827e3e43e56b
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit f85794de8572e9f89137da8a3cba9b33438afd34)