Support neutron-{api,rpc-server} services
In Debian, all is now Python 3. In such world, neutron-server doesn't
work over SSL, because of Eventlet bugs/limitations. So instead of
running neutron-server, we have to run neutron-api on top of mod_wsgi or
on top of uwsgi. In such a setup, we need to run neutron-rpc-server, as
the Neutron WSGI application doesn't do the AMQP jobs.

Change-Id: I4926327c513e2c79f8fbe1563fa6a7e70683401b