Add support for dynamic routing bpg dragent
This patch adds support to install and configure the
bgp dynamic routing agent. It also install the
dynamic-routing package since that is required by
the bgp dragent.

These packages already exists on Debian and Ubuntu but
since it was removed from RPM packages when dynamic-routing
was moved out of the Neutron tree there are work to add it here [1].
So until [1] is merged and hopefully backported to Queens for the latest
stable release this will work propery on RedHat based OS.

[1] https://review.rdoproject.org/r/#/c/7935/

Change-Id: Id8a97b337eff68860547e07e675b79bde94628ba
(cherry picked from commit 75a999369a7d5dc7f11680274863e3efa38e40db)