Add support for n-server to manage n-d-r package
When enabling 'bgp' in the service_plugins neutron-server
will try to start and load the bgp service plugin but if
the neutron-dynamic-routing package is not there it will
fail to start.

This adds a mechanism to enable installation of the n-d-r
packages by setting the enable_dr_package in neutron::server
same as done for VPNaaS, LBaaS etc.

Change-Id: I8a50f36ea3e824ebd959e627e0de90cd28ec4e56
(cherry picked from commit ef6d25c2fd47f1e925c073dd5681312f506b5759)