Set metadata_agent_package to false for RedHat
The metadata_agent_package variable in the params.pp
is not set on RedHat but only on Debian based OS family.

This causes an warning because the if statement in [1] [2].

We should set the variable to false to suppress the
warning message.

[1] https://github.com/openstack/puppet-neutron/blob/master/manifests/agents/metadata.pp#L122
[2] Unknown variable: '::neutron::params::metadata_agent_package'

Change-Id: I4dc1c0867ebc5a61865aabe46aa05716bd4e7452
(cherry picked from commit 8d3ffb584d06ea1b4fdcce7896485339cb0baf69)