Deprecate all Neutron LBaaS code
The Neutron LBaaS project was deprecated in Queens and will
be removed in a future release.

Please see here for all information about the deprecation [1].

The Octavia project [2] was borne out of Neutron LBaaS and replaces
all it's functionality and currently supports integration with
Neutron LBaaS as well to allow a migration between the two solutions.

[1] https://wiki.openstack.org/wiki/Neutron/LBaaS/Deprecation
[2] https://docs.openstack.org/octavia/latest/reference/introduction.html

Change-Id: Ic3899f1366cad935545e9567f774f74f228b5800