Remove br-tun check from beaker
Currently this check is flapping because the ovs agent is hitting
timeouts when trying to get a response from the neutron-server. This is
likely because we start all the services at the same time and if
neutron-server takes too long to start as well as the exponential
timeout calculations from the ovs agent being to long, the br-tun check
will fail.  Let's comment this out until we can tune the timeouts better
or address the performance issues.

Change-Id: I8158e674775922549eda24115fae0099c740c419