Deprecate firewal_driver option in the Neutron's ML2 config
Config option ``firewall_driver`` should be set in the Neutron agent's, not
in the server.
Usage of this option in the Neutron server was there just for backward
compatibility with old agents which can't report what driver they are using.
Since Newton all Neutron drivers are reporting that in heartbeat messages and
there is no need to keep configure this in the neutron server's side.

Neutron is removing this option from the neutron server in patch:
https://review.opendev.org/#/c/726351/

Change-Id: I19a2aa4151cd798fa35248bf3f39dd0819fdc8af
(cherry picked from commit 1868cb6dc5fb25e492bfc08742b7fee4c9b5680b)