Update configuration for ovs integration bridge
The DEFAULT/ovs_integration_bridge parameter was deprecated in neutron
in favor of the new ovs/integration_bridge[1] parameter, so
puppet-neutron should use the new parameter accordingly.

Also, that parameter is also available in l3 agent configuration, so
the new parameter, ovs_integration_bridge, was added to l3 agent
manifest.

[1] https://github.com/openstack/neutron/commit/33fb446add87a3427a246444b03befc46adb385d

Change-Id: Ia558d97754aa3ad0bb00630b965b31e281c4b9f0