Ensure resource_provider_bandwitdhs is unset by default
Currently puppet-neutron supports the resource_provider_bandwidths
parameter for ovs agent and sriov agent, but it doesn't provide the way
to clear existing settings.

This change makes sure that the resource_provider_bandwiths parameter
is unset by default so that operators can reset the parameter to
the default value defined in neutron service itself.

Change-Id: Icd834f4f4ca6578c836935614931af2a3e79bf7d