Fix L3 agent's extensions list
L3 agent's extensions list should be passed as an array
to the neutron::agents::l3 class and then it should be converted
to the comma separated string in the L3 agent's config.
This patch changes expected type of the variable and
adds that such conversion from an array to string.

Change-Id: I0650085d2aa2c3492b175c565976705c977fd806