Deprecate the useless quota_packet_filter parameter
The quota_packet_filter parameter is no longer used in core neutron.
The parameter was used by NEC OpenFlow plugin but the plugin was
migrated to an independent repository[1][2].

Currently puppet-neutron doesn't support the plugin so the parameter
is no longer used in a deployment managed by puppet-neutron.

Because we haven't seen any interest in adding support for the plugin
and the plugin itself has not been updated for a while, let's deprecate
the useless parameter so that we can remove it in the next cycle.

[1] https://opendev.org/x/networking-nec
[2] cfa8f537710d8d7ef407b4194295f15ed03b3fa0

Change-Id: I355526f41f8ec17498dc79a5849f90e5385ccc4e